Move test config files to build-system/test-configs/
