+++ build-system/tasks/changelog.js
@@ -27,7 +27,7 @@ const assert = require('assert');
 const BBPromise = require('bluebird');
 const childProcess = require('child_process');
 const colors = require('ansi-colors');
-const config = require('../config');
+const config = require('../test-configs/config');
 const extend = require('util')._extend;
 const git = require('gulp-git');
 const log = require('fancy-log');
