+++ build-system/tasks/dev-dashboard-tests.js
@@ -15,7 +15,7 @@
  */
 'use strict';
 
-const config = require('../config');
+const config = require('../test-configs/config');
 const deglob = require('globs-to-files');
 const Mocha = require('mocha');
 const {isTravisBuild} = require('../travis');
