+++ build-system/tasks/e2e/index.js
@@ -18,7 +18,7 @@
 
 const argv = require('minimist')(process.argv.slice(2));
 const ciReporter = require('../mocha-ci-reporter');
-const config = require('../../config');
+const config = require('../../test-configs/config');
 const glob = require('glob');
 const log = require('fancy-log');
 const Mocha = require('mocha');
