+++ build-system/tasks/todos.js
@@ -20,7 +20,7 @@ const colors = require('ansi-colors');
 const gulp = require('gulp');
 const log = require('fancy-log');
 const request = BBPromise.promisify(require('request'));
-const srcGlobs = require('../config').presubmitGlobs;
+const srcGlobs = require('../test-configs/config').presubmitGlobs;
 const through2 = require('through2');
 
 const {GITHUB_ACCESS_TOKEN} = process.env;
